Updating wait check for startup due to race condition.
